# pkg/swagger/docs/swagger.yaml
definitions:
  emplacc-api_api_v1_dto.AttendanceResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.Attendance'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
  emplacc-api_api_v1_dto.AttendancesByUserResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.AttendancesByUser'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
  emplacc-api_api_v1_dto.AttendancesListResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.AttendancesPage'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  emplacc-api_api_v1_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorDetail'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto.ResponseMeta:
    properties:
      pagination:
        $ref: '#/definitions/emplacc-api_api_v1_dto.Pagination'
      trace_id:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_request.CreateAttendance:
    properties:
      actual_start:
        type: string
      code_reviews:
        type: integer
      commits:
        type: integer
      date:
        type: string
      end_work:
        type: string
      merge_requests:
        type: integer
      planned_start:
        type: string
      status:
        type: string
      user_id:
        type: string
      workday_hours:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto_request.UpdateAttendance:
    properties:
      actual_start:
        type: string
      code_reviews:
        type: integer
      commits:
        type: integer
      date:
        type: string
      end_work:
        type: string
      merge_requests:
        type: integer
      planned_start:
        type: string
      status:
        type: string
      workday_hours:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto_response.Attendance:
    properties:
      actual_start:
        type: string
      code_reviews:
        type: integer
      commits:
        type: integer
      created_at:
        type: string
      date:
        type: string
      end_work:
        type: string
      id:
        type: string
      merge_requests:
        type: integer
      planned_start:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      workday_hours:
        type: integer
    type: object
  emplacc-api_api_v1_dto_response.AttendancesByUser:
    properties:
      attendances:
        items:
          $ref: '#/definitions/emplacc-api_api_v1_dto_response.Attendance'
        type: array
      user_id:
        type: string
    type: object
  emplacc-api_api_v1_dto_response.AttendancesPage:
    properties:
      attendances:
        items:
          $ref: '#/definitions/emplacc-api_api_v1_dto_response.Attendance'
        type: array
    type: object
...
info:
  contact: {}
paths:
  /attendance:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of attendances
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.AttendancesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Attendances
      tags:
      - Attendance
    post:
      consumes:
      - application/json
      description: Create a new attendance record
      parameters:
      - description: Attendance data
        in: body
        name: attendance
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.CreateAttendance'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.AttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Create Attendance
      tags:
      - Attendance
  /attendance/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an attendance record by its ID
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Delete Attendance
      tags:
      - Attendance
    get:
      consumes:
      - application/json
      description: Retrieve a specific attendance by its ID
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.AttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Get Attendance
      tags:
      - Attendance
    patch:
      consumes:
      - application/json
      description: Update an existing attendance record
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated attendance data
        in: body
        name: attendance
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.UpdateAttendance'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.AttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Update Attendance
      tags:
      - Attendance
...
  /users/{user_id}/attendance:
    get:
      consumes:
      - application/json
      description: Retrieve a list of attendances for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.AttendancesByUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Attendances by User
      tags:
      - Attendance
...
swagger: "2.0"